Refactor tests strategy and add tests files for resync
