+++ Videos/subtitles/spongebob.srt
@@ -1,0 +1,90 @@
+1
+00:00:00,918 --> 00:00:02,294
+[alarm clock sounding]
+
+2
+00:00:02,377 --> 00:00:03,754
+Time for boating school.
+
+3
+00:00:04,046 --> 00:00:05,214
+Let her rip, Gary.
+
+4
+00:00:12,054 --> 00:00:13,138
+Ah.
+
+5
+00:00:19,019 --> 00:00:20,354
+Ta-da!
+
+6
+00:00:20,437 --> 00:00:22,523
+Not bad, partner in crime.
+
+7
+00:00:22,773 --> 00:00:24,983
+The toast could have been
+a little darker though.
+
+8
+00:00:25,108 --> 00:00:26,109
+See you later.
+
+9
+00:00:26,193 --> 00:00:27,277
+Hey SpongeBob.
+
+10
+00:00:27,361 --> 00:00:28,820
+Wanna go jellyfishing?
+
+11
+00:00:28,946 --> 00:00:31,240
+Sorry Patrick, I can't.
+I have school today.
+
+12
+00:00:31,365 --> 00:00:34,451
+What am I supposed to do all
+day while you're at school?
+
+13
+00:00:34,576 --> 00:00:36,995
+I don't know.
+What do you normally do when I'm gone?
+
+14
+00:00:37,120 --> 00:00:39,373
+Wait for you to get back.
+
+15
+00:00:39,957 --> 00:00:43,794
+Wait a second, Patrick.
+Why don't you come to school with me?
+
+16
+00:00:43,919 --> 00:00:45,838
+Hey, that's a great idea.
+
+17
+00:00:45,963 --> 00:00:49,633
+You and me in school together
+as classmates. Think about it!
+
+18
+00:00:53,262 --> 00:00:54,304
+Wow.
+
+19
+00:00:54,888 --> 00:00:58,976
+♪ Road, road, road, road ♪
+
+20
+00:00:59,101 --> 00:01:00,435
+♪ Road, road ♪
+
+21
+00:01:00,561 --> 00:01:03,355
+♪ When I'm on the road ♪
+
